Relocate SAP sending related code
Moved SAP sending related code under directory "send" to be more
consistent with fetching related code being under directory "fetch".
